Сonverted the area into a fractional number. Сonverted Participant into interface. Removed null in competition method. Removed instance of in obstacles

+++ ua.ithillel.homework/srс/ua/ithillel/lesson7/figurearea/Main.java
@@ -7,11 +7,11 @@ public class Main {
 
     }
 
-    public static int allFigureArea(CalculateFigureArea[] figures) {
+    public static double allFigureArea(CalculateFigureArea[] figures) {
         double result = 0;
         for (int i = 0; i < figures.length; i++) {
             result += figures[i].calculate();
         }
-        return (int) result;
+        return result;
     }
 }
